\name{netpca-package}
\alias{netpca-package}
\docType{package}
\title{GSEA for Principal Components Analysis}
\description{Package for identifying reproducible principal components across datasets and characterizing those components by GSEA.}
\details{
\tabular{ll}{Package: \tab netpca\cr
Maintainer: \tab Princy Parsana <princyparsana@jhu.edu>\cr
Depends: \tab R (>= 2.10), limma\cr
Imports: \tab BiocGenerics\cr
Suggests: \tab pensim\cr
Author: \tab Princy Parsana and Levi Waldron\cr
Version: \tab 0.2\cr
Date: \tab 2012-09-29\cr
License: \tab GPL (>2.0)\cr
Title: \tab GSEA for Principal Components Analysis\cr}
}
\author{Princy Parsana and Levi Waldron}

\keyword{ package }


